// api/v2/types.ts
export interface UserLite {
  id: number;
  username: string;
  slug: string;
  avatar_url: string;
  is_staff: boolean;
  is_superuser: boolean;
}
// ...
export interface Comment {
  id: number;
  gallery_id: number;
  poster: UserLite;
  post_date: number;
  body: string;
}